requirements.txt
click==8.1.3
Flask==2.2.3
importlib-metadata==6.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
numpy==1.24.2
packaging==23.0
Pillow==9.5.0
PyMuPDF==1.21.1
pytesseract==0.3.10
scikit-learn==1.2.2
scipy==1.10.1
threadpoolctl==3.1.0
tqdm==4.65.0
Werkzeug==2.2.3
zipp==3.15.0
